requirements.txt
# Python 3.9
allure-behave==2.8.36
allure-python-commons==2.8.36
attrs==20.3.0
Authlib==0.15.3
aws-kinesis-agg==1.1.6
beautifulsoup4==4.9.3
behave==1.2.6
boto3==1.17.32
botocore==1.20.32
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
chromedriver-autoinstaller==0.2.2
cryptography==3.4.6
cx-Oracle==8.3.0
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
Faker==6.6.2
fastjsonschema==2.15.0
ffmpeg-python==0.2.0
flatten-dict==0.3.0
future==0.18.2
fuzzywuzzy==0.18.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
h11==0.12.0
hpack==3.0.0
httpcore==0.12.3
httplib2==0.19.0
httpx==0.17.1
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
jmespath==0.10.0
jsonpath-ng==1.5.2
kiwisolver==1.4.0
lxml==4.8.0
MarkupSafe==2.0.0
matplotlib==3.5.1
mss==5.1.0
mysql-connector-python==8.0.23
networkx==2.5
numpy==1.21.5
O365==2.0.14
oauthlib==3.1.0
openpyxl==3.0.7
opencv-python==4.5.5.64
packaging==20.9
pandas==1.4.1
parse==1.19.0
parse-type==0.5.2
pathlib2==2.3.5
Pillow==9.0.1
pluggy==0.13.1
ply==3.11
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.3.0
requests==2.25.1
requests-oauthlib==1.3.0
retry==0.9.2
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.3.6
scikit-image==0.19.2
scipy==1.8.0
selenium==3.141.0
simple-salesforce==1.10.1
six==1.15.0
sniffio==1.2.0
soupsieve==2.2
SQLAlchemy==1.4.1
stringcase==1.2.0
text-unidecode==1.3
tifffile==2021.3.17
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
pdfminer==20191125
jinja2==2.10.1
xlrd==2.0.1
Levenshtein==0.18.1
py-expression-eval==0.3.14